reorganize and board viewer js
Moves it into it's own module
