+++ web/static/js/battle_snake/index.js
@@ -1,0 +1,1 @@
+export {default as BoardViewer} from "./board_viewer";
